+++ src/queryParameters/index.ts
@@ -1,3 +1,4 @@
+export * from './extractBaseFieldSensitivityClassificationsParameter';
 export * from './extractChangemakerParameters';
 export * from './extractCreatedByParameters';
 export * from './extractFunderParameters';